+++ crypto-sodium/lib/Crypto/Encrypt/Public.hs
@@ -39,6 +39,7 @@ module Crypto.Encrypt.Public
   , toSecretKey
   , keypair
   , keypairFromSeed
+  , unsafeKeypairFromSeed
 
   -- * Nonce
   , Nonce
@@ -50,15 +51,16 @@ module Crypto.Encrypt.Public
   ) where
 
 import Data.ByteArray (ByteArray, ByteArrayAccess, ScrubbedBytes, withByteArray)
-import Data.ByteArray.Sized as Sized (alloc, allocRet)
+import Data.ByteArray.Sized as Sized (SizedByteArray, alloc, allocRet)
 import Data.ByteString (ByteString)
 import Data.Functor (void)
 import Data.Proxy (Proxy(..))
-import NaCl.Box
-  (Nonce, PublicKey, SecretKey, keypair, toNonce, toPublicKey, toSecretKey)
+import System.IO.Unsafe (unsafeDupablePerformIO)
 
 import qualified Libsodium as Na
 
+import NaCl.Box
+  (Nonce, PublicKey, SecretKey, keypair, toNonce, toPublicKey, toSecretKey)
 import qualified NaCl.Box as NaCl.Box
 
 
@@ -121,10 +123,22 @@ decrypt
 decrypt = NaCl.Box.open
 
 
+-- | Seed for deterministically generating a keypair.
+--
+-- In accordance with Libsodium's documentation, the seed must be of size
+-- @Na.CRYPTO_BOX_SEEDBYTES@.
+--
+-- This type is parametrised by the actual data type that contains
+-- bytes. This can be, for example, a @ByteString@.
+type Seed a = SizedByteArray Na.CRYPTO_BOX_SEEDBYTES a
+
+
 -- | Generate a new 'SecretKey' together with its 'PublicKey' from a given seed.
+--
+-- The given seed must be @Na.CRYPTO_BOX_SEEDBYTES@ bytes long.
 keypairFromSeed
   :: ByteArrayAccess seed
-  => seed
+  => Seed seed
   -> IO (PublicKey ByteString, SecretKey ScrubbedBytes)
 keypairFromSeed seed = do
   allocRet Proxy $ \skPtr ->
@@ -132,3 +146,13 @@ keypairFromSeed seed = do
     withByteArray seed $ \sdPtr ->
     -- always returns 0, so we don’t check it
     void $ Na.crypto_box_seed_keypair pkPtr skPtr sdPtr
+
+-- | Generate a new 'SecretKey' together with its 'PublicKey' from a given seed,
+-- in a pure context.
+--
+-- The given seed must be @Na.CRYPTO_BOX_SEEDBYTES@ bytes long.
+unsafeKeypairFromSeed
+  :: ByteArrayAccess seed
+  => Seed seed
+  -> (PublicKey ByteString, SecretKey ScrubbedBytes)
+unsafeKeypairFromSeed = unsafeDupablePerformIO . keypairFromSeed